V1 tests were screwing up with v2 tests.  I am moving them into their own file
