+++ test/recaptcha_test.rb
@@ -19,23 +19,12 @@ class RecaptchaClientHelperTest < Minitest::Test
     end
   end
 
-  def test_recaptcha_api_version_default
-    assert_equal(Recaptcha.configuration.api_version, Recaptcha::RECAPTCHA_API_VERSION)
-  end
-
   def test_recaptcha_tags_v2
+    Recaptcha.configuration.api_version = 'v2'
     # match a v2 only tag
     assert_match /data-sitekey/, recaptcha_tags
     # refute a v1 only tag
     refute_match /\/challenge\?/, recaptcha_tags
-  end
-
-  def test_recaptcah_tags_v1
-    Recaptcha.configuration.api_version = 'v1'
-    # match a v1 only tag
-    assert_match /\/challenge\?/, recaptcha_tags
-    # refute a v2 only tag
-    refute_match /data-sitekey/, recaptcha_tags
   end
 
   def test_ssl_by_default
@@ -61,30 +50,5 @@ class RecaptchaClientHelperTest < Minitest::Test
       Recaptcha.configuration.public_key = nil
       recaptcha_tags
     end
-  end
-
-  def test_different_configuration_within_with_configuration_block
-    key = Recaptcha.with_configuration(:public_key => '12345') do
-      Recaptcha.configuration.public_key
-    end
-
-    assert_equal('12345', key)
-  end
-
-  def test_v1_with_api_v1?
-    Recaptcha.configuration.api_version = 'v1'
-    assert Recaptcha.configuration.v1?
-    refute Recaptcha.configuration.v2?
-  end
-
-  def test_v2_with_api_v2?
-    Recaptcha.configuration.api_version = 'v2'
-    assert Recaptcha.configuration.v2?
-    refute Recaptcha.configuration.v1?
-  end
-
-  def test_reset_configuration_after_with_configuration_block
-    Recaptcha.with_configuration(:public_key => '12345')
-    assert_equal('0000000000000000000000000000000000000000', Recaptcha.configuration.public_key)
   end
 end
